mathjax & fix exporting bug

+++ MarkDownEditor/Model/DocumentExporter.cs
@@ -16,6 +16,7 @@ namespace MarkDownEditor.Model
             {
                 { "Plain Html", new PlainHTMLExporter()},
                 { "Html", new HTMLExporter()},
+                { "Html Local Mathjax", new HTMLWithLocalMathJaxExporter()},
                 { "RTF", new RFTExporter()},
                 { "Docx", new DocxExporter()},
                 { "Epub", new EpubExporter()},
@@ -71,10 +72,44 @@ namespace MarkDownEditor.Model
 
             Process process = new Process();
             process.StartInfo.FileName = "pandoc";
+            string mathjax = Properties.Settings.Default.ShowMathJax ? "--mathjax" : "";
             process.StartInfo.Arguments = 
                 cssFile == null 
-                ? $"\"{sourceCodePath}\" -f {markdownType} -t html --ascii -s -o \"{outputPath}\""
-                : $"\"{sourceCodePath}\" -f {markdownType} -t html --ascii -s -H {tmpFile} -o \"{outputPath}\"";
+                ? $"\"{sourceCodePath}\" -f {markdownType} -t html {mathjax} --ascii -s -o \"{outputPath}\""
+                : $"\"{sourceCodePath}\" -f {markdownType} -t html {mathjax} --ascii -s -H {tmpFile} -o \"{outputPath}\"";
+            process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
+            process.Start();
+            process.WaitForExit();
+
+            File.Delete(tmpFile);
+        }
+    }
+
+    public class HTMLWithLocalMathJaxExporter : IDocumentExporter
+    {
+        string mathjaxJsFilePath = new Uri(Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "MathJax", "MathJax.js")).AbsoluteUri + "?config=TEX_CHTML";
+        public void Export(string markdownType, string sourceCodePath, string cssFile, string outputPath)
+        {
+            var tmpFile = Path.GetTempFileName();
+            if (cssFile != null)
+            {
+                StreamReader sr = new StreamReader(cssFile);
+                var cssContent = sr.ReadToEnd();
+                sr.Close();
+                StreamWriter sw = new StreamWriter(tmpFile);
+                sw.WriteLine("<style type=\"text/css\">");
+                sw.WriteLine(cssContent);
+                sw.WriteLine("</style>");
+                sw.Close();
+            }
+
+            Process process = new Process();
+            process.StartInfo.FileName = "pandoc";
+            string mathjax = Properties.Settings.Default.ShowMathJax ? $"--mathjax={mathjaxJsFilePath}" : "";
+            process.StartInfo.Arguments =
+                cssFile == null
+                ? $"\"{sourceCodePath}\" -f {markdownType} -t html {mathjax} --ascii -s -o \"{outputPath}\""
+                : $"\"{sourceCodePath}\" -f {markdownType} -t html {mathjax} --ascii -s -H {tmpFile} -o \"{outputPath}\"";
             process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
             process.Start();
             process.WaitForExit();
@@ -143,15 +178,18 @@ namespace MarkDownEditor.Model
 
             DocumentExporter.Export("Html", markdownType, cssFile, sourceCodePath, tmpFilePath);
 
-            File.WriteAllBytes(outputPath, HtmlToXConverter.ConvertToPdf(File.ReadAllText(tmpFilePath)));
-            //Process process = new Process();
-            //process.StartInfo.FileName = "wkhtmltopdf";
-            //process.StartInfo.Arguments = $"\"{tmpFilePath}\" \"{outputPath}\"";
-            //process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
-            //process.Start();
-            //process.WaitForExit();
-            //if (process.ExitCode != 0)
-            //    throw new Exception(Properties.Resources.FailedToExport + "\n" + "wkhtmltopdf error" + process.ExitCode);
+            if (File.Exists(outputPath))
+                File.Delete(outputPath);
+
+            //File.WriteAllBytes(outputPath, HtmlToXConverter.ConvertToPdf(File.ReadAllText(tmpFilePath)));
+            Process process = new Process();
+            process.StartInfo.FileName = "WkHtmlToPdfWrapper";
+            process.StartInfo.Arguments = $"\"{tmpFilePath}\" \"{outputPath}\"";
+            process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
+            process.Start();
+            process.WaitForExit();
+            if (process.ExitCode != 0)
+                throw new Exception(Properties.Resources.FailedToExport + "\n" + "wkhtmltopdf error" + process.ExitCode);
 
             File.Delete(tmpFilePath);
         }
@@ -165,19 +203,22 @@ namespace MarkDownEditor.Model
 
             DocumentExporter.Export("Html", markdownType, cssFile, sourceCodePath, tmpFilePath);
 
-            //Process process = new Process();
-            //process.StartInfo.FileName = "wkhtmltoimage";
-            //process.StartInfo.Arguments = $"--width 600 \"{tmpFilePath}\" \"{outputPath}\"";
-            //process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
-            //process.Start();
-            //process.WaitForExit();
-            //if (process.ExitCode != 0)
-            //    throw new Exception(Properties.Resources.FailedToExport + "\n" + "wkhtmltoimage error" + process.ExitCode);
+            if (File.Exists(outputPath))
+                File.Delete(outputPath);
 
-            var html = File.ReadAllText(tmpFilePath);
-            var extension = Path.GetExtension(outputPath).Remove(0, 1).ToLower();
-            var image = HtmlToXConverter.ConvertToImage(html, extension, 600, 0);
-            File.WriteAllBytes(outputPath, image);
+            Process process = new Process();
+            process.StartInfo.FileName = "WkHtmlToImageWrapper";
+            process.StartInfo.Arguments = $"\"{tmpFilePath}\" \"{outputPath}\"";
+            process.StartInfo.WindowStyle = ProcessWindowStyle.Hidden;
+            process.Start();
+            process.WaitForExit();
+            if (process.ExitCode != 0)
+                throw new Exception(Properties.Resources.FailedToExport + "\n" + "wkhtmltoimage error" + process.ExitCode);
+
+            //var html = File.ReadAllText(tmpFilePath);
+            //var extension = Path.GetExtension(outputPath).Remove(0, 1).ToLower();
+            //var image = HtmlToXConverter.ConvertToImage(html, extension, 600, 0);
+            //File.WriteAllBytes(outputPath, image);
 
             File.Delete(tmpFilePath);
         }
